% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_app.R
\docType{data}
\name{nhs_app}
\alias{nhs_app}
\title{NHS digital contact tracing activity}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 137 rows and 3 columns.
}
\source{
\url{https://www.gov.uk/government/publications/nhs-covid-19-app-statistics}

Originally licensed under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government Licence v3.0}
}
\usage{
data("nhs_app")
}
\description{
Summary data collected as part of the NHS digital contact tracing app
monitoring. This describes the number of alerts issued, and venue "check-ins".
}
\details{
\describe{
\item{ \code{date} (date) }{
The date
}
\item{ \code{alerts} (int) }{
Number of alerts
}
\item{ \code{visits} (int) }{
Number of check-ins
}
}
}
\concept{datasets}
\keyword{datasets}
