% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{timeline}
\alias{timeline}
\title{Timeline of events}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 3 columns.
}
\source{
\url{https://en.wikipedia.org/wiki/Timeline_of_the_COVID-19_pandemic_in_the_United_Kingdom}
}
\usage{
data("timeline")
}
\description{
Major events in the UK COVID-19 pandemic, limited to lock-downs, vaccination
roll-out and first identification of major variants.
}
\details{
\describe{
\item{ \code{label} (chr) }{ The event }
\item{ \code{start} (date) }{ The start date }
\item{ \code{end} (date) }{ The end date if a period }
}
}
\examples{
dplyr::glimpse(timeline)
}
\concept{datasets}
\keyword{datasets}
