% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_population_2019.R
\docType{data}
\name{uk_population_2019}
\alias{uk_population_2019}
\title{Country, regional, and subnational total population estimates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 398 rows and 4 columns.
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates}

Originally licensed under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government Licence v3.0}
}
\usage{
data("uk_population_2019")
}
\description{
ONS National and subnational mid-year population estimates for the UK and its
constituent countries by administrative area, age and sex (including
components of population change, median age and population density).
}
\details{
Mid-2019: April 2019 local authority district codes edition of this dataset.
This is UK wide and covers country, regions and LTLA (2019 boundaries)
\subsection{\code{uk_population_2019} dataframe with 398 rows and 4 columns}{

\describe{
\item{ \code{name} (chr) }{
The region name
}
\item{ \code{code} (chr) }{
The region code
}
\item{ \code{codeType} (chr) }{
The ONS geographical region code type (including year)
}
\item{ \code{population} (dbl) }{
the count of the population in that age group
}
}
}
}
\examples{
dplyr::glimpse(uk_population_2019)
}
\concept{datasets}
\keyword{datasets}
