% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukc19-package.R
\docType{package}
\name{ukc19-package}
\alias{ukc19}
\alias{ukc19-package}
\title{ukc19: Datasets from the UK COVID-19 outbreak}
\description{
Provides simple access to a small selection of pre-wrangled data sets relevant to the COVId-19 outbreak in the UK for teaching and demo purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ai4ci.github.io/ukc19}
  \item \url{https://github.com/ai4ci/ukc19}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
